requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0            # SQLite async driver
asyncpg>=0.29.0              # PostgreSQL async driver (اختياري)

# ─── Telegram ─────────────────────────────────────────────────────────────────
pyrogram>=2.0.106
tgcrypto>=1.2.5              # تشفير أسرع لـ Pyrogram

# ─── AI Agents ────────────────────────────────────────────────────────────────
crewai>=0.30.0
langchain>=0.2.0
langchain-openai>=0.1.6
langchain-community>=0.2.0
openai>=1.30.0
anthropic>=0.28.0

# ─── Vector Memory ────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Utilities ────────────────────────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.1
pandas>=2.2.0
httpx>=0.27.0
